+++ config/response-code.js
@@ -1,0 +1,10 @@
+'use strict';
+module.exports = {
+  204: 'Reset Content',
+  205: 'No Content',
+  400: 'Bad Request',
+  403: 'Forbidden',
+  404: 'Page Not Found',
+  500: 'Internal Server Error',
+  502: 'Bad Gateway',
+};