+++ src/utils/response.js
@@ -1,4 +1,14 @@
-('use strict');
+'use strict';
+const responseCode = require('../../config/response-code');
+
+function errPage(res, status, msg) {
+  let statusCode = status;
+  const messageCode = responseCode[statusCode];
+  const message = [{ code: `${statusCode}`, title: `${messageCode}`, message: `${msg || ''}` }];
+  res.set('Cache-Control', 'private, no-cache');
+  res.status(statusCode).render('error-page', { message });
+  return errPage;
+}
 
 const json = (res, data) => {
   res.json({
@@ -7,36 +17,4 @@ const json = (res, data) => {
   });
 };
 
-const error500 = (res, error, status = 500) => {
-  res.status(status).json({
-    status: false,
-    code: `-${status}`,
-    error: `Internal Server Error`,
-    message: `${error}`,
-  });
-};
-
-const error502 = (res, error, status = 502) => {
-  const message = [
-    {
-      code: status,
-      title: 'Bad Gateway',
-      message: `NomCDN either was unable to connect to <code><span>${error}</span></code> or received an invalid response from <code><span>${error}</span></code><br />Please try your request again in a moment.`,
-    },
-  ];
-  res.set('Cache-Control', 'private, no-cache');
-  res.status(status).render('error', { message });
-};
-
-const error403 = (res, error, status = 403) => {
-  const message = [{ code: status, title: 'Forbidden', message: `${error}` }];
-  res.set('Cache-Control', 'private, no-cache');
-  res.status(status).render('error', { message });
-};
-
-module.exports = {
-  json,
-  error500,
-  error502,
-  error403,
-};
+module.exports = { errPage, json };